# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My-Trip API Documentation",
        "description": "API documentation for My-Trip Platform",
        "contact": {
            "email": "support@mytrip.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/my-trip/public",
            "description": "Development Server"
        },
        {
            "url": "https://blueviolet-hummingbird-437500.hostingersite.com/public",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/flights/search": {
            "post": {
                "tags": [
                    "Flights"
                ],
                "summary": "البحث عن الرحلات (Search for flights)",
                "description": "البحث عن رحلات الطيران المتاحة عبر Travelopro API مع عرض كافة التفاصيل والأسعار. (Search for flight availability using Travelopro API).",
                "operationId": "searchFlights",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "لغة الاستجابة (ar, en)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "en",
                            "enum": [
                                "en",
                                "ar"
                            ]
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "journeyType",
                                    "OriginDestinationInfo",
                                    "class",
                                    "adults"
                                ],
                                "properties": {
                                    "journeyType": {
                                        "type": "string",
                                        "enum": [
                                            "OneWay",
                                            "Return",
                                            "Circle"
                                        ],
                                        "example": "OneWay"
                                    },
                                    "OriginDestinationInfo": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "departureDate",
                                                "airportOriginCode",
                                                "airportDestinationCode"
                                            ],
                                            "properties": {
                                                "departureDate": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2024-12-01"
                                                },
                                                "returnDate": {
                                                    "description": "Required if journeyType is Return",
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2024-12-10"
                                                },
                                                "airportOriginCode": {
                                                    "type": "string",
                                                    "example": "DXB"
                                                },
                                                "airportDestinationCode": {
                                                    "type": "string",
                                                    "example": "LHR"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "class": {
                                        "type": "string",
                                        "enum": [
                                            "First",
                                            "Business",
                                            "Economy",
                                            "PremiumEconomy"
                                        ],
                                        "example": "Economy"
                                    },
                                    "adults": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "childs": {
                                        "type": "integer",
                                        "example": 0
                                    },
                                    "infants": {
                                        "type": "integer",
                                        "example": 0
                                    },
                                    "requiredCurrency": {
                                        "type": "string",
                                        "example": "SAR"
                                    },
                                    "airlineCode": {
                                        "description": "Two letter code of the preferred airline",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "directFlight": {
                                        "description": "True for only direct flights, false for all (or 1/0)",
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "تم استرجاع الرحلات بنجاح",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Flights retrieved successfully."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "خطأ في التحقق من البيانات",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/flights/airports": {
            "get": {
                "tags": [
                    "Flights"
                ],
                "summary": "Get list of airports",
                "description": "Retrieve a list of supported airports from Travelopro.",
                "operationId": "getAirports",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "The language of the response (ar, en)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "en",
                            "enum": [
                                "en",
                                "ar"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful retrieval",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Airports retrieved successfully."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "AirportCode": {
                                                        "type": "string",
                                                        "example": "DXB"
                                                    },
                                                    "AirportName": {
                                                        "type": "string",
                                                        "example": "Dubai International Airport"
                                                    },
                                                    "City": {
                                                        "type": "string",
                                                        "example": "Dubai"
                                                    },
                                                    "Country": {
                                                        "type": "string",
                                                        "example": "United Arab Emirates"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/flights/airlines": {
            "get": {
                "tags": [
                    "Flights"
                ],
                "summary": "Get list of airlines",
                "description": "Retrieve a list of supported airlines from Travelopro.",
                "operationId": "getAirlines",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "The language of the response (ar, en)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "en",
                            "enum": [
                                "en",
                                "ar"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful retrieval",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Airlines retrieved successfully."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "AirLineCode": {
                                                        "type": "string",
                                                        "example": "EK"
                                                    },
                                                    "AirLineName": {
                                                        "type": "string",
                                                        "example": "Emirates"
                                                    },
                                                    "AirLineLogo": {
                                                        "type": "string",
                                                        "example": "https://travelnext.works/api/airlines/EK.gif"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/flights/validate-fare": {
            "post": {
                "tags": [
                    "Flights"
                ],
                "summary": "التحقق من صحة السعر (Validate flight fare)",
                "description": "التحقق مما إذا كان سعر الرحلة المحدد لا يزال متاحاً وصحيحاً قبل البدء بعملية الحجز. (Verify if the selected flight fare is still available).",
                "operationId": "validateFare",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "لغة الاستجابة (ar, en)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "en",
                            "enum": [
                                "en",
                                "ar"
                            ]
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "session_id",
                                    "fare_source_code"
                                ],
                                "properties": {
                                    "session_id": {
                                        "type": "string",
                                        "example": "7906efba-09db-4481-8c60-0d7f5b5e6c44"
                                    },
                                    "fare_source_code": {
                                        "type": "string",
                                        "example": "MTY2ODE2Njg2Ml8yNjA5Mzk"
                                    },
                                    "fare_source_code_inbound": {
                                        "description": "Required for Indian Domestic RoundTrip",
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "تم التحقق من السعر بنجاح",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Fare is valid."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "السعر غير متاح أو خطأ في البيانات",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Fare is no longer valid or available."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use a token to access protected routes",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}